change pos of mouse event to short type, or when mouse is out of control, x or y will be more than 65000 change scroll range to total width, or it will act wired

+++ nui/ui/NFrameBase.h
@@ -138,9 +138,9 @@ namespace nui
             virtual void SetLayoutable(bool layoutable);
             virtual bool IsLayoutable() const;
 
-            virtual void OnMouseDown(int x, int y);
+            virtual void OnMouseDown(short x, short y);
             virtual void OnMouseUp();
-            virtual void OnMouseMove(int x, int y);
+            virtual void OnMouseMove(short x, short y);
             virtual void OnMouseHover();
             virtual void OnMouseLeave();
